third_party/blink/web_tests/platform/generic/fast/forms/input-user-select-none-expected.txt

PASS sample.value is "wxy"
PASS sample.value is "wxy"
PASS successfullyParsed is true

TEST COMPLETE
